fix subtle bugs
in IE10, if you use window.open() and try to access pageXOffset in the
new window, it fails.  You must use document.body.scrollTop/scrollLeft.

Tests were failing intermittently because the DOM modification was
screwing with asynchronous tests.  So, on some platforms, the tests are
run in a new window.  Except mobile.

Fix the list of IE browsers for sub-testing